pyproject.toml
[tool.poetry]
name = "is-alive"
version = "0.1.0"
description = "Periodically check website availability"
authors = ["Eduard Maghakyan <edi.maghakyan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
kafka-python = "^2.0.2"
pinject = "^0.14.1"
environs = "^9.3.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest-watch = "^4.2.0"
mypy = "^0.812"
flake8 = "^3.9.2"
isort = "^5.8.0"
black = "^21.5b1"
safety = "^1.10.3"
bandit = "^1.7.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
